requirements.txt
absl-py==0.7.1
chainer==6.1.0
cycler==0.10.0
filelock==3.0.12
future==0.17.1
grpcio==1.22.0
joblib==0.13.2
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
numpy==1.16.4
Pillow==6.0.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tb-nightly==1.15.0a20190711
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
typing==3.6.6
typing-extensions==3.6.6
Werkzeug==0.15.4
